+++ test/models/gator_config_test.dart
@@ -6,6 +6,13 @@ import '../test_helpers.dart';
 
 void main() {
   group('GatorConfig', () {
+    test('const constructor', () {
+      expect(
+        const GatorConfig(className: '', colors: [], outputPath: ''),
+        equals(const GatorConfig(className: '', colors: [], outputPath: '')),
+      );
+    });
+
     test('constructor assertion', () {
       expect(
         () => GatorConfig.fromYaml(YamlMap()),